Fix blog related-articles query crashing every /blog/[slug] page load
getRelatedByCategory's ORDER BY used a bare sql`0` as a no-op tiebreaker
when sportingEventId is null (the case for every travel_craft article,
which never carries one) — Postgres treats a literal integer in ORDER BY
as a positional reference, not a constant expression, and column position
0 doesn't exist. Cast to sql`0::int` so it's a real constant.

Confirmed live in Sentry: "ORDER BY position 0 is not in select list",
24 events / 17 users over the last day, first seen on release 18748914781f.
Verified fix against production DB directly.

+++ lib/queries/blog.ts
@@ -88,7 +88,7 @@ export async function getRelatedByCategory(
     .orderBy(
       sportingEventId
         ? sql`CASE WHEN ${blogArticles.sportingEventId} = ${sportingEventId} THEN 0 ELSE 1 END`
-        : sql`0`,
+        : sql`0::int`,
       sql`CASE WHEN ${blogArticles.sport} && ARRAY[${sql.join(sport.map((s) => sql`${s}`), sql`, `)}]::sport[] THEN 0 ELSE 1 END`,
       sql`array_length(${blogArticles.sport}, 1) asc`,
       desc(blogArticles.publishedAt),
